requirements.txt
jieba
synonyms
torchtext
pytorch >= 1.5.0
